requirements.txt
################################################################
#
# Python package requirements for the project:
#
# If you are using Anaconda then you can install all required
# Python packages by running the following commands in a shell:
#
#     conda create --name tf python=3.6
#     source activate tf
#     pip install -r requirements.txt
#
# Note that you have to edit this file to select whether you
# want to install the CPU or GPU version of TensorFlow.
#
################################################################
# Basic packages used in many of the tutorials.

numpy
scipy
matplotlib
Pillow
scikit-learn
# jupyter
# pandas   

################################################################
# TensorFlow can be installed either as CPU or GPU versions.
# You select which one to install by (un)commenting these lines.

# tensorflow		# CPU Version of TensorFlow.
tensorflow-gpu==1.12.0	# GPU version of TensorFlow.

################################################################
